requirement.txt
torch
fairscale
fire
sentencepiece
llama-cpp-python